Refactor

+++ theme/gatsby-node.js
@@ -1,4 +1,7 @@
 const fs = require('fs-extra')
+const path = require('path')
+
+const pack = require('./package.json')
 
 const createHome = require('./create/createHome')
 const createPost = require('./create/createPost')
@@ -6,9 +9,10 @@ const createTags = require('./create/createTags')
 
 exports.onPreBootstrap = ({ reporter }) => {
   const contentFolder = 'theme-content'
-  const themeContent = `../node_modules/gatsby-theme-minimal/${contentFolder}`
+  const isThemePath = fs.existsSync(path.resolve(`node_modules/${pack.name}`))
+  const themeContent = path.resolve(`node_modules/${pack.name}/${contentFolder}`)
 
-  if (!fs.existsSync(contentFolder)) {
+  if (!fs.existsSync(contentFolder) && isThemePath) {
     reporter.info(`creating the ${contentFolder} directory`)
     fs.copySync(themeContent, contentFolder)
   }
